update eslint config and move more live-feed content to subcomponents
