+++ app/components/custom/settings-form.tsx
@@ -9,16 +9,11 @@ import {
   SelectValue,
 } from '@/app/components/ui/select';
 import { Button } from '@/app/components/ui/button';
-import { Input } from '@/app/components/ui/input';
 import { 
   getDefaultClient, 
   setDefaultClient, 
-  getSnapchainNodeUrl, 
-  setSnapchainNodeUrl,
-  validateSnapchainUrl,
   type DefaultClient 
 } from '../../lib/local-storage';
-import { SNAPCHAIN_NODE_BASE_URL } from '../../lib/utils';
 
 interface SettingsFormProps {
   onClose?: () => void;
@@ -26,43 +21,15 @@ interface SettingsFormProps {
 
 export function SettingsForm({ onClose }: SettingsFormProps) {
   const [defaultClientState, setDefaultClientState] = useState<DefaultClient>('farcaster');
-  const [snapchainUrl, setSnapchainUrl] = useState(SNAPCHAIN_NODE_BASE_URL);
-  const [urlError, setUrlError] = useState('');
-  const [isValidating, setIsValidating] = useState(false);
   const [mounted, setMounted] = useState(false);
 
   useEffect(() => {
     setMounted(true);
     setDefaultClientState(getDefaultClient());
-    setSnapchainUrl(getSnapchainNodeUrl());
   }, []);
 
-  const validateUrl = async (url: string) => {
-    if (!url.trim()) {
-      setUrlError('URL is required');
-      return false;
-    }
-
-    setIsValidating(true);
-    setUrlError('');
-    
-    const isValid = await validateSnapchainUrl(url);
-    
-    if (!isValid) {
-      setUrlError('Invalid URL or /v1/info endpoint not accessible');
-    }
-    
-    setIsValidating(false);
-    return isValid;
-  };
-
-  const handleSave = async () => {
-    const isUrlValid = await validateUrl(snapchainUrl);
-    
-    if (!isUrlValid) return;
-    
+  const handleSave = () => {
     setDefaultClient(defaultClientState);
-    setSnapchainNodeUrl(snapchainUrl);
     onClose?.();
   };
 
@@ -93,26 +60,11 @@ export function SettingsForm({ onClose }: SettingsFormProps) {
         </Select>
       </div>
       
-      {/* <div className="space-y-2">
-        <label className="text-sm font-medium">snapchain node url</label>
-        <Input
-          value={snapchainUrl}
-          onChange={(e) => {
-            setSnapchainUrl(e.target.value);
-            setUrlError('');
-          }}
-          placeholder={SNAPCHAIN_NODE_BASE_URL}
-          className="border border-black rounded-none bg-white hover:bg-gray-50"
-        />
-        {urlError && <p className="text-sm text-red-600">{urlError}</p>}
-      </div> */}
-      
       <Button 
         onClick={handleSave}
-        disabled={isValidating}
-        className="border border-black bg-white text-black hover:bg-gray-50 rounded-none font-medium disabled:opacity-50"
+        className="border border-black bg-white text-black hover:bg-gray-50 rounded-none font-medium"
       >
-        {isValidating ? 'validating...' : 'save'}
+        save
       </Button>
     </div>
   );
