+++ app/lib/local-storage.ts
@@ -1,10 +1,7 @@
 'use client';
 
-import { SNAPCHAIN_NODE_BASE_URL } from './utils';
-
 export const LOCAL_STORAGE_KEYS = {
   DEFAULT_CLIENT: 'default_client',
-  SNAPCHAIN_NODE_BASE_URL: 'SNAPCHAIN_NODE_BASE_URL',
 } as const;
 
 export type DefaultClient = 'farcaster' | 'baseapp';
@@ -19,37 +16,4 @@ export const setDefaultClient = (client: DefaultClient): void => {
   if (typeof window === 'undefined') return;
   localStorage.setItem(LOCAL_STORAGE_KEYS.DEFAULT_CLIENT, client);
   window.dispatchEvent(new Event('localStorageChange'));
-};
-
-export const getSnapchainNodeUrl = (): string => {
-  if (typeof window === 'undefined') return SNAPCHAIN_NODE_BASE_URL;
-  return localStorage.getItem(LOCAL_STORAGE_KEYS.SNAPCHAIN_NODE_BASE_URL) || SNAPCHAIN_NODE_BASE_URL;
-};
-
-export const setSnapchainNodeUrl = (url: string): void => {
-  if (typeof window === 'undefined') return;
-  localStorage.setItem(LOCAL_STORAGE_KEYS.SNAPCHAIN_NODE_BASE_URL, url);
-  window.dispatchEvent(new Event('localStorageChange'));
-};
-
-export const validateSnapchainUrl = async (url: string): Promise<boolean> => {
-  try {
-    new URL(url);
-  } catch {
-    return false;
-  }
-
-  try {
-    let snapchainHttpUrl = url;
-    if (snapchainHttpUrl === SNAPCHAIN_NODE_BASE_URL) {
-      snapchainHttpUrl += ':3381';
-    }
-    const response = await fetch(`${snapchainHttpUrl}/v1/info`, { 
-      method: 'GET',
-      signal: AbortSignal.timeout(5000)
-    });
-    return response.ok;
-  } catch {
-    return false;
-  }
 };
