+++ eslint.config.mjs
@@ -1,0 +1,13 @@
+import { FlatCompat } from '@eslint/eslintrc'
+
+const compat = new FlatCompat({
+  baseDirectory: import.meta.dirname,
+})
+
+const eslintConfig = [{
+  ignores: ["node_modules/**", ".next/**", "out/**", "build/**", "next-env.d.ts"]
+}, ...compat.config({
+  extends: ['next/core-web-vitals', 'prettier'],
+})]
+
+export default eslintConfig
